수정 : 파일 업로드 속도 개선 작업

+++ src/apis/board/board.resolver.ts
@@ -1,7 +1,5 @@
-import { CACHE_MANAGER, Inject, UseGuards } from '@nestjs/common';
-import { ElasticsearchService } from '@nestjs/elasticsearch';
-import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { Cache } from 'cache-manager';
+import { Ip, UseGuards } from '@nestjs/common';
+import { Args, Context, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GraphQLJSONObject } from 'graphql-type-json';
 import { GqlAuthAccessGuard } from 'src/commons/auth/gql-auth.guard';
 import { CurrentUser, ICurrentUser } from 'src/commons/auth/gql-user-param';
@@ -9,7 +7,14 @@ import { BoardService } from './board.service';
 import { BoardTagsInput } from './dto/boardTags.input';
 import { CreateBoardInput } from './dto/createBoard.input';
 import { UpdateBoardInput } from './dto/updateBoard.input';
-import { Board, Tags } from './entities/board.entity';
+import {
+  AGE_GROUP_ENUM,
+  Board,
+  GENDER_ENUM,
+  SUB_CATEGORY_NAME_ENUM,
+} from './entities/board.entity';
+import { RealIP } from 'nestjs-real-ip';
+import { IpAddress } from 'src/commons/ip/ip';
 
 @Resolver()
 export class BoardResolver {
@@ -17,68 +22,54 @@ export class BoardResolver {
     //
 
     private readonly boardService: BoardService,
-    private readonly elasticsearchService: ElasticsearchService,
-    @Inject(CACHE_MANAGER)
-    private readonly cacheManager: Cache,
   ) {}
+
+  @Query(() => [Board])
+  async fetchBoardTitle(
+    //
+    @Args('title') title: string,
+  ) {
+    return this.boardService.test({ title });
+  }
 
   @Query(() => GraphQLJSONObject)
   async fetchBoardWithTags(
     //
-    @Args('tags') tags: Tags,
+    @Args({ name: 'tags', type: () => [String] }) tags: string[],
   ) {
-    const tagsData = tags.names.reduce((acc, cur) => {
-      const tag = cur.substring(1);
-      return acc === '' ? acc + tag : acc + ' ' + tag;
-    }, '');
-
-    const redisInData = await this.cacheManager.get(tagsData);
-
-    if (redisInData) {
-      return redisInData;
-    } else {
-      const data = await this.elasticsearchService.search({
-        index: 'board',
-        size: 10000,
-        sort: 'sortData',
-        _source: [
-          'boardtitle',
-          'boardwriter',
-          'boardlikecount',
-          'boardhit',
-          'createat',
-        ],
-        query: {
-          match: {
-            tags: {
-              query: tagsData,
-              operator: 'and',
-            },
-          },
-        },
-      });
-
-      await this.cacheManager.set(tagsData, data, { ttl: 2 });
-      return data;
-    }
+    return this.boardService.elasticsearchFindTags({ tags });
   }
 
   @Query(() => Board)
-  fetchBoard(@Args('boardId') boardId: string) {
-    return this.boardService.findOne({ boardId });
+  fetchBoard(
+    //
+    @Args('boardId') boardId: string,
+    @Context() context: any,
+  ) {
+    const ip = context.req.clientIp;
+    console.log(ip);
+
+    return this.boardService.findOne({ boardId, ip });
   }
 
   @Query(() => [Board])
-  fetchTestBoards(
+  fetchBoards() {
+    return this.boardService.findAll();
+  }
+
+  @Query(() => [Board])
+  fetchBoardCategoryPick(
     //
-    @Args('boardTagsInput') boardTagsInput: BoardTagsInput,
+    @Args({ name: 'category', type: () => SUB_CATEGORY_NAME_ENUM })
+    category: SUB_CATEGORY_NAME_ENUM,
   ) {
-    return this.boardService.findTest({ boardTagsInput });
+    return this.boardService.findPickList({ category });
   }
 
   @Query(() => [Board])
   fetchGenderBoards(
-    @Args('gender') gender: string,
+    @Args({ name: 'gender', type: () => GENDER_ENUM })
+    gender: GENDER_ENUM,
     @Args({ name: 'page', type: () => Int }) page: number,
   ) {
     return this.boardService.findGender({ gender, page });
@@ -86,7 +77,8 @@ export class BoardResolver {
 
   @Query(() => [Board])
   fetchAgeGroupBoards(
-    @Args('ageGroup') ageGroup: string,
+    @Args({ name: 'ageGroup', type: () => AGE_GROUP_ENUM })
+    ageGroup: AGE_GROUP_ENUM,
     @Args({ name: 'page', type: () => Int }) page: number,
   ) {
     return this.boardService.findAgeGroup({ ageGroup, page });
@@ -94,35 +86,27 @@ export class BoardResolver {
 
   @Query(() => [Board])
   fetchAgeGroupWithGenderBoards(
-    @Args('gender') gender: string,
-    @Args('ageGroup') ageGroup: string,
+    @Args({ name: 'gender', type: () => GENDER_ENUM })
+    gender: GENDER_ENUM,
+    @Args({ name: 'ageGroup', type: () => AGE_GROUP_ENUM })
+    ageGroup: AGE_GROUP_ENUM,
     @Args({ name: 'page', type: () => Int }) page: number,
   ) {
     return this.boardService.findGenderWithAgeGroup({ gender, ageGroup, page });
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Board)
   createBoard(
     //
-
+    @CurrentUser() currentUser: ICurrentUser,
     @Args('createBoardInput') createBoardInput: CreateBoardInput,
     @Args('boardTagsInput') boardTagsInput: BoardTagsInput,
   ) {
-    return this.boardService.create({ createBoardInput, boardTagsInput }); // 받아온 name 넘기기 service로
-  }
-
-  @UseGuards(GqlAuthAccessGuard)
-  @Mutation(() => Board)
-  createBoardUserLoggedin(
-    //
-    @CurrentUser() currentUser: ICurrentUser,
-    @Args('createBoardInput') createBoardInput: CreateBoardInput,
-    @Args('createBoardTagsInput') boardTagsInput: BoardTagsInput,
-  ) {
-    return this.boardService.loginCreate({
-      currentUser,
+    return this.boardService.create({
       createBoardInput,
       boardTagsInput,
+      currentUser,
     });
   }
 
@@ -138,8 +122,12 @@ export class BoardResolver {
     });
   }
 
+  @UseGuards(GqlAuthAccessGuard)
   @Mutation(() => Boolean)
-  deleteBoard(@Args('boardId') boardId: string) {
-    return this.boardService.delete({ boardId });
+  deleteBoard(
+    @CurrentUser() currentUser: ICurrentUser,
+    @Args('boardId') boardId: string,
+  ) {
+    return this.boardService.delete({ boardId, currentUser });
   }
 }
